@@ -1,45 +1,14 @@
-import Vue from 'vue'
-import App from './App.vue'
-/*VueRouter*/
-import router from  './router'
-import hljs from 'highlight.js'
-Vue.use(hljs.vuePlugin)
-import './assets/css/index.css'
+import { createApp } from "vue";
+import { createPinia } from "pinia";
 
-import { library } from '@fortawesome/fontawesome-svg-core'
-import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
-/* Icons */
-import {
-  faRocket,
-  faDotCircle,
-  faWindowRestore,
-  faEdit,
-  faCalendarAlt,
-  faEnvelope,
-  faTasks,
-  faWindowMaximize,
-  faChartBar,
-  faChevronDown,
-  faCode,
-  faAngleDown,
-  faAngleUp,
-  faLayerGroup,
-  faGripHorizontal,
-  faAngleLeft,
-  faClock,
-  faRetweet,
-  faTags,
-  faTv,
-  faPlug,
-  faTable, faShoppingBag, faLiraSign, faTrash, faPlusCircle, faAngleDoubleRight, faInfo
-} from '@fortawesome/free-solid-svg-icons'
-library.add(faRocket, faDotCircle, faWindowRestore, faEdit, faCalendarAlt,faEnvelope,faTasks,faWindowMaximize,faChartBar,faChevronDown,faCode,faAngleDown,faAngleUp,faChartBar,faLayerGroup,faGripHorizontal,faAngleLeft,faClock,faRetweet,faTags,faTv,faPlug,faTable,faShoppingBag,faLiraSign,faTrash,faPlusCircle,faAngleDoubleRight,faInfo)
+import App from "./App.vue";
+import router from "./router";
 
-Vue.component('font-awesome-icon', FontAwesomeIcon)
+import "./assets/main.css";
 
-Vue.config.productionTip = false
+const app = createApp(App);
 
-new Vue({
-  router,
-  render: h => h(App),
-}).$mount('#app')
+app.use(createPinia());
+app.use(router);
+
+app.mount("#app");
